Check if the resource exists before deleting it

+++ controllers/mysqlcluster_controller.go
@@ -163,7 +163,7 @@ func (r *MySQLClusterReconciler) reconcileV1(ctx context.Context, req ctrl.Reque
 
 		r.ClusterManager.Stop(req.NamespacedName)
 
-		if err := r.finalizeV1(ctx, req, cluster); err != nil {
+		if err := r.finalizeV1(ctx, cluster); err != nil {
 			log.Error(err, "failed to finalize")
 			return ctrl.Result{}, err
 		}
@@ -1127,21 +1127,27 @@ func (r *MySQLClusterReconciler) reconcileV1RestoreJob(ctx context.Context, req 
 	return nil
 }
 
-func (r *MySQLClusterReconciler) finalizeV1(ctx context.Context, req ctrl.Request, cluster *mocov1beta1.MySQLCluster) error {
+func (r *MySQLClusterReconciler) finalizeV1(ctx context.Context, cluster *mocov1beta1.MySQLCluster) error {
 	secretName := cluster.ControllerSecretName()
 	secret := &corev1.Secret{}
-	secret.SetNamespace(r.SystemNamespace)
-	secret.SetName(secretName)
-	if err := r.Delete(ctx, secret); err != nil {
-		return fmt.Errorf("failed to delete controller secret %s: %w", secretName, err)
+	err := r.Get(ctx, client.ObjectKey{Namespace: r.SystemNamespace, Name: secretName}, secret)
+	if err == nil {
+		if err := r.Delete(ctx, secret); err != nil {
+			return fmt.Errorf("failed to delete controller secret %s: %w", secretName, err)
+		}
+	} else if !apierrors.IsNotFound(err) {
+		return err
 	}
 
 	certName := cluster.CertificateName()
 	cert := certificateObj.DeepCopy()
-	cert.SetNamespace(r.SystemNamespace)
-	cert.SetName(certName)
-	if err := r.Delete(ctx, cert); err != nil {
-		return fmt.Errorf("failed to delete certificate %s: %w", certName, err)
+	err = r.Get(ctx, client.ObjectKey{Namespace: r.SystemNamespace, Name: certName}, cert)
+	if err == nil {
+		if err := r.Delete(ctx, cert); err != nil {
+			return fmt.Errorf("failed to delete certificate %s: %w", certName, err)
+		}
+	} else if !apierrors.IsNotFound(err) {
+		return err
 	}
 
 	return nil
